+++ CHANGELOG.md
@@ -1,5 +1,19 @@
-# Revision history for doxygen-parser
+# Revision history for `doxygen-parser`
+
+## ?.?.? -- YYYY-mm-dd
+
+### Breaking changes
+
+### New features
+
+### Minor changes
+
+### Bug fixes
 
 ## 0.1.0 -- Unreleased
 
-* First version. Extracted from hs-bindgen.
+* First version. Extracted from
+  [`hs-bindgen`](https://github.com/well-typed/hs-bindgen). See
+  [hs-bindgen#1884][hs-bindgen-is-1884].
+
+[hs-bindgen-is-1884]: https://github.com/well-typed/hs-bindgen/issues/1884